requirements.txt
openai>=1.95.1
python-dotenv>=0.19.0
openai-agents>=0.1.0